format & refactoring
