+++ packages/editor/src/editor-integration/create-renderers.tsx
@@ -7,6 +7,7 @@ import { ArticleStaticRenderer } from '@editor/plugins/article/static'
 import { BlanksExerciseStaticRenderer } from '@editor/plugins/blanks-exercise/static'
 import { BoxStaticRenderer } from '@editor/plugins/box/static'
 import { DropzoneImageStaticRenderer } from '@editor/plugins/dropzone-image/static'
+import { EdusharingAssetStaticRenderer } from '@editor/plugins/edusharing-asset/static'
 import { EquationsStaticRenderer } from '@editor/plugins/equations/static'
 import { ExerciseStaticRenderer } from '@editor/plugins/exercise/static'
 import { GeogebraStaticRenderer } from '@editor/plugins/geogebra/static'
@@ -16,6 +17,7 @@ import { InputExerciseStaticRenderer } from '@editor/plugins/input-exercise/stat
 import { MultimediaStaticRenderer } from '@editor/plugins/multimedia/static'
 import { RowsStaticRenderer } from '@editor/plugins/rows/static'
 import { ScMcExerciseStaticRenderer } from '@editor/plugins/sc-mc-exercise/static'
+import { SerloInjectionStaticRenderer } from '@editor/plugins/serlo-injection/static'
 import { SerloTableStaticRenderer } from '@editor/plugins/serlo-table/static'
 import { GenericContentTypeStaticRenderer } from '@editor/plugins/serlo-template-plugins/generic-content/static'
 import { StaticSolutionRenderer } from '@editor/plugins/solution/static'
@@ -28,8 +30,6 @@ import { VideoStaticRenderer } from '@editor/plugins/video/static'
 import { EditorPluginType } from '@editor/types/editor-plugin-type'
 import { TemplatePluginType } from '@editor/types/template-plugin-type'
 import { ComponentProps } from 'react'
-import { EdusharingAssetStaticRenderer } from '@editor/plugins/edusharing-asset/static'
-import { SerloInjectionStaticRenderer } from '@editor/plugins/serlo-injection/static'
 
 export function createRenderers(): InitRenderersArgs {
   return {
