+++ packages/editor/src/package/editor.tsx
@@ -3,17 +3,17 @@ import { createBasicPlugins } from '@editor/editor-integration/create-basic-plug
 import { createRenderers } from '@editor/editor-integration/create-renderers'
 import { editorPlugins } from '@editor/plugin/helpers/editor-plugins'
 import { editorRenderers } from '@editor/plugin/helpers/editor-renderer'
+import { LtikContext } from '@editor/plugins/edusharing-asset/ltik-context'
+import { EditorPluginType } from '@editor/types/editor-plugin-type'
 import { SupportedLanguage } from '@editor/types/language-data'
-import { TemplatePluginType } from '@editor/types/template-plugin-type.js'
-import { LtikContext } from '@editor/plugins/edusharing-asset/static.jsx'
-import { InstanceDataProvider } from '@/contexts/instance-context'
-import { LoggedInDataProvider } from '@/contexts/logged-in-data-context'
-import '@/assets-webkit/styles/serlo-tailwind.css'
+import { TemplatePluginType } from '@editor/types/template-plugin-type'
 import React from 'react'
 
-import { defaultSerloEditorProps } from './config.js'
-import { editorData } from './editor-data.js'
-import { EditorPluginType } from './index.js'
+import '@/assets-webkit/styles/serlo-tailwind.css'
+import { defaultSerloEditorProps } from './config'
+import { editorData } from './editor-data'
+import { InstanceDataProvider } from '@/contexts/instance-context'
+import { LoggedInDataProvider } from '@/contexts/logged-in-data-context'
 
 export interface SerloEditorProps {
   children: EditorProps['children']
@@ -51,7 +51,6 @@ export function SerloEditor(props: SerloEditorProps) {
   return (
     <InstanceDataProvider value={instanceData}>
       <LoggedInDataProvider value={loggedInData}>
-        {/* TODO: Find solution to get ltik into static renderer without context */}
         <LtikContext.Provider value={_ltik}>
           <div className="serlo-editor-hacks">
             <Editor initialState={initialState} onChange={onChange}>
