+++ packages/editor/src/package/serlo-renderer.tsx
@@ -1,5 +1,6 @@
 import { createRenderers } from '@editor/editor-integration/create-renderers'
 import { editorRenderers } from '@editor/plugin/helpers/editor-renderer'
+import { LtikContext } from '@editor/plugins/edusharing-asset/ltik-context'
 import { StaticRenderer } from '@editor/static-renderer/static-renderer'
 import type { AnyEditorDocument } from '@editor/types/editor-plugins'
 import type { SupportedLanguage } from '@editor/types/language-data'
@@ -8,7 +9,6 @@ import { defaultSerloEditorProps } from './config'
 import { editorData } from './editor-data'
 import { InstanceDataProvider } from '@/contexts/instance-context'
 import { LoggedInDataProvider } from '@/contexts/logged-in-data-context'
-import { LtikContext } from '@editor/plugins/edusharing-asset/static'
 
 export interface SerloRendererProps {
   language?: SupportedLanguage
@@ -30,7 +30,6 @@ export function SerloRenderer(props: SerloRendererProps) {
   return (
     <InstanceDataProvider value={instanceData}>
       <LoggedInDataProvider value={loggedInData}>
-        {/* TODO: Find solution to get ltik into static renderer without context */}
         <LtikContext.Provider value={_ltik}>
           <div className="serlo-editor-hacks">
             <StaticRenderer {...props} />
