+++ packages/editor/src/plugins/edusharing-asset/ltik-context.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const LtikContext = createContext<string | undefined>(undefined)
