+++ packages/editor/src/plugins/serlo-injection/static.tsx
@@ -1,4 +1,5 @@
 import { EditorInjectionDocument } from '@editor/types/editor-plugins'
+
 import { SerloInjectionRenderer } from './renderer'
 
 export function SerloInjectionStaticRenderer(props: EditorInjectionDocument) {
